+++ controller/credentialcontroller.php
@@ -11,10 +11,9 @@
 
 namespace OCA\Passman\Controller;
 
-use OCA\Files_External\NotFoundException;
 use OCA\Passman\Db\SharingACL;
+use OCA\Passman\Service\SettingsService;
 use OCA\Passman\Utility\NotFoundJSONResponse;
-use OCP\AppFramework\Db\DoesNotExistException;
 use OCP\AppFramework\Http;
 use OCP\AppFramework\Http\DataResponse;
 use OCP\IRequest;
@@ -25,7 +24,7 @@ use OCA\Passman\Activity;
 use OCA\Passman\Service\ActivityService;
 use OCA\Passman\Service\CredentialRevisionService;
 use OCA\Passman\Service\ShareService;
-use OCP\IUser;
+
 
 class CredentialController extends ApiController {
 	private $userId;
@@ -33,6 +32,7 @@ class CredentialController extends ApiController {
 	private $activityService;
 	private $credentialRevisionService;
 	private $sharingService;
+	private $settings;
 
 	public function __construct($AppName,
 								IRequest $request,
@@ -40,7 +40,8 @@ class CredentialController extends ApiController {
 								CredentialService $credentialService,
 								ActivityService $activityService,
 								CredentialRevisionService $credentialRevisionService,
-								ShareService $sharingService
+								ShareService $sharingService,
+								SettingsService $settings
 	) {
 		parent::__construct($AppName, $request);
 		$this->userId = $userId;
@@ -48,7 +49,9 @@ class CredentialController extends ApiController {
 		$this->activityService = $activityService;
 		$this->credentialRevisionService = $credentialRevisionService;
 		$this->sharingService = $sharingService;
+		$this->settings = $settings;
 	}
+
 
 	/**
 	 * @NoAdminRequired
@@ -85,7 +88,7 @@ class CredentialController extends ApiController {
 		);
 		$credential = $this->credentialService->createCredential($credential);
 		$link = ''; // @TODO create direct link to credential
-		if(!$credential->getHidden()) {
+		if (!$credential->getHidden()) {
 			$this->activityService->add(
 				Activity::SUBJECT_ITEM_CREATED_SELF, array($label, $this->userId),
 				'', array(),
@@ -139,14 +142,18 @@ class CredentialController extends ApiController {
 		);
 
 
-		if ($storedCredential->getUserId() !== $this->userId) {
+		if (!hash_equals($storedCredential->getUserId(), $this->userId)) {
 			$acl = $this->sharingService->getCredentialAclForUser($this->userId, $storedCredential->getGuid());
 			if ($acl->hasPermission(SharingACL::WRITE)) {
 				$credential['shared_key'] = $storedCredential->getSharedKey();
 			} else {
 				return new DataResponse(['msg' => 'Not authorized'], Http::STATUS_UNAUTHORIZED);
 			}
+			if ($this->settings->isEnabled('user_sharing_enabled')) {
+				return new DataResponse(['msg' => 'Not authorized'], Http::STATUS_UNAUTHORIZED);
+			}
 		}
+
 		$link = ''; // @TODO create direct link to credential
 		if ($revision_created) {
 			$activity = 'item_apply_revision';
@@ -154,13 +161,13 @@ class CredentialController extends ApiController {
 				$activity . '_self', array($label, $this->userId, $revision_created),
 				'', array(),
 				$link, $this->userId, Activity::TYPE_ITEM_ACTION);
-		} else if (($storedCredential->getDeleteTime() === 0) && (int) $delete_time > 0) {
+		} else if (($storedCredential->getDeleteTime() === 0) && (int)$delete_time > 0) {
 			$activity = 'item_deleted';
 			$this->activityService->add(
 				$activity . '_self', array($label, $this->userId),
 				'', array(),
 				$link, $this->userId, Activity::TYPE_ITEM_ACTION);
-		} else if (($storedCredential->getDeleteTime() > 0) && (int) $delete_time === 0) {
+		} else if (($storedCredential->getDeleteTime() > 0) && (int)$delete_time === 0) {
 			$activity = 'item_recovered';
 			$this->activityService->add(
 				$activity . '_self', array($label, $this->userId),
@@ -183,7 +190,7 @@ class CredentialController extends ApiController {
 
 		try {
 			$acl_list = $this->sharingService->getCredentialAclList($storedCredential->getGuid());
-		} catch (DoesNotExistException $exception) {
+		} catch (\Exception $exception) {
 			// Just check if we have an acl list
 		}
 		if (!empty($acl_list)) {
@@ -204,7 +211,7 @@ class CredentialController extends ApiController {
 
 			foreach ($acl_list as $sharingACL) {
 				$target_user = $sharingACL->getUserId();
-				if($target_user === $this->userId){
+				if ($target_user === $this->userId) {
 					continue;
 				}
 				$this->activityService->add(
@@ -212,22 +219,22 @@ class CredentialController extends ApiController {
 					'', array(),
 					$link, $target_user, Activity::TYPE_ITEM_ACTION);
 			}
-			if ($this->userId !== $storedCredential->getUserId()) {
+			if (!hash_equals($this->userId, $storedCredential->getUserId())) {
 				$this->activityService->add(
 					$activity, $params,
 					'', array(),
 					$link, $storedCredential->getUserId(), Activity::TYPE_ITEM_ACTION);
 			}
 		}
-		if($set_share_key === true){
+		if ($set_share_key === true) {
 			$storedCredential->setSharedKey($shared_key);
 			$credential['shared_key'] = $shared_key;
 		}
-		if($unshare_action === true){
+		if ($unshare_action === true) {
 			$storedCredential->setSharedKey('');
 			$credential['shared_key'] = '';
 		}
-		if(!$skip_revision) {
+		if (!$skip_revision) {
 			$this->credentialRevisionService->createRevision($storedCredential, $storedCredential->getUserId(), $credential_id, $this->userId);
 		}
 		$credential = $this->credentialService->updateCredential($credential);
@@ -259,26 +266,23 @@ class CredentialController extends ApiController {
 	 * @NoCSRFRequired
 	 */
 	public function getRevision($credential_guid) {
-	    try {
-            $credential = $this->credentialService->getCredentialByGUID($credential_guid);
-        }
-        catch (DoesNotExistException $ex){
-            return new NotFoundJSONResponse();
-        }
+		try {
+			$credential = $this->credentialService->getCredentialByGUID($credential_guid);
+		} catch (\Exception $ex) {
+			return new NotFoundJSONResponse();
+		}
 
-        // If the request was made by the owner of the credential
-        if ($this->userId === $credential->getUserId()) {
-            $result = $this->credentialRevisionService->getRevisions($credential->getId(), $this->userId);
-        }
-        else {
-            $acl = $this->sharingService->getACL($this->userId, $credential_guid);
-            if ($acl->hasPermission(SharingACL::HISTORY)){
-                $result = $this->credentialRevisionService->getRevisions($credential->getId());
-            }
-            else {
-                return new NotFoundJSONResponse();
-            }
-        }
+		// If the request was made by the owner of the credential
+		if ($this->userId === $credential->getUserId()) {
+			$result = $this->credentialRevisionService->getRevisions($credential->getId(), $this->userId);
+		} else {
+			$acl = $this->sharingService->getACL($this->userId, $credential_guid);
+			if ($acl->hasPermission(SharingACL::HISTORY)) {
+				$result = $this->credentialRevisionService->getRevisions($credential->getId());
+			} else {
+				return new NotFoundJSONResponse();
+			}
+		}
 
 		return new JSONResponse($result);
 	}
@@ -296,17 +300,17 @@ class CredentialController extends ApiController {
 	 * @NoAdminRequired
 	 * @NoCSRFRequired
 	 */
-	public function updateRevision($credential_guid, $revision_id, $credential_data){
+	public function updateRevision($credential_guid, $revision_id, $credential_data) {
 		$revision = null;
 		try {
 			$this->credentialService->getCredentialByGUID($credential_guid, $this->userId);
-		} catch (DoesNotExistException $e) {
+		} catch (\Exception $e) {
 			return new NotFoundJSONResponse();
 		}
 
-		try{
+		try {
 			$revision = $this->credentialRevisionService->getRevision($revision_id);
-		} catch(DoesNotExistException $exception){
+		} catch (\Exception $exception) {
 			return new NotFoundJSONResponse();
 		}