+++ CHANGELOG.md
@@ -1,6 +1,20 @@
-owncloud-passman (0.0.1)
-* **Security**: Security description here
-* **Backwards incompatible change**: Changes in the API
-* **New dependency**: New dependencies such as a new ownCloud or PHP version
-* **Bugfix**: Bugfix description
-* **Enhancement**: New feature description
+## [Unreleased]
+### Updated
+- Updated to passman 2.0.0
+
+## 2.0.0 – 2016-12-31
+### Added
+- Password sharing
+- Vaults
+- Change vault passwords
+- Unit tests
+
+### Changed
+- Passman API overhaul
+- Rewrite of code base
+- New passman repo at https://github.com/nextcloud/passman
+### Fixed
+- A lot of small bug fixes
+
+### Removed
+- Old passman API